net/ipv6: convert skip_notify_on_dev_down sysctl to u8
JIRA: https://issues.redhat.com/browse/RHEL-33410
Upstream Status: net.git commit ef62c0ae6db11c095880e473db9f846132d7eba8
Conflicts:
 - include/net/netns/ipv6.h: mismatch because of missing commit 9ee11f0fff20 ("ipv6:
   ioam: Data plane support for Pre-allocated Trace")

commit ef62c0ae6db11c095880e473db9f846132d7eba8
Author: Eric Dumazet <edumazet@google.com>
Date:   Thu Jun 1 16:04:45 2023 +0000

    net/ipv6: convert skip_notify_on_dev_down sysctl to u8

    Save a bit a space, and could help future sysctls to
    use the same pattern.

    Acked-by: Matthieu Baerts <matthieu.baerts@tessares.net>

+++ include/net/netns/ipv6.h
@@ -51,7 +51,7 @@ struct netns_sysctl_ipv6 {
 	int max_dst_opts_len;
 	int max_hbh_opts_len;
 	int seg6_flowlabel;
-	int skip_notify_on_dev_down;
+	u8 skip_notify_on_dev_down;
 	u8 fib_notify_on_flag_change;
 };
